Add name recognition
When user joins, their name is stored in the map to keep the name reserved and to retrieve the nickname when sending messages to all the users. When user disconnects their key is getting deleted from the map, so that other users could login with that name. But it may be confusing taking over user names.