Implement basic rooms functionality
Make rooms to be created via the form. Once the form is submitted user gets redirected to the new room they created with the id received from the server. Make rooms list to be updated to all users when new room is created or when someone left the room. Remove react strict mode to have expected component unmount functionality.